+++ fluidSimLib/gridFractionalSinglePhasePressureSolver3.h
@@ -19,7 +19,7 @@ public:
    
     void solve(
         const faceCenteredGrid3& input,
-        double timeIntervalInSeconds,
+        FloatType timeIntervalInSeconds,
         faceCenteredGrid3* output,
         const scalarField3& boundarySdf
         = constantScalarField3(mathUtil::maxFloat()),
