+++ fluidSimLib/gridPointGenerator3.h
@@ -10,7 +10,7 @@ public:
    
     void forEachPoint(
         const boundingBox3& boundingBox,
-        double spacing,
+        FloatType spacing,
         const std::function<bool(const vector3&)>& callback) const;
 };
